kb-admin/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1
mysql-connector-python==8.2.0
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.3
langchain-anthropic==0.1.1
python-multipart==0.0.6
aiofiles==23.2.1
alembic==1.13.1
sqlalchemy==2.0.25
httpx==0.26.0
